Fix for handling 2gb+ files.
git-svn-id: http://svn.vuze.com/public/client@906 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/core3/disk/impl/PieceMapEntry.java
@@ -10,10 +10,10 @@ package org.gudy.azureus2.core3.disk.impl;
 // More readable and simpler as a separate class
 public class PieceMapEntry {
 	private DiskManagerFileInfoImpl _file;
-	private int _offset;
+	private long _offset;
 	private int _length;
 
-	public PieceMapEntry(DiskManagerFileInfoImpl file, int offset, int length) {
+	public PieceMapEntry(DiskManagerFileInfoImpl file, long offset, int length) {
 		_file = file;
 		_offset = offset;
 		_length = length;
@@ -24,7 +24,7 @@ public class PieceMapEntry {
 	public void setFile(DiskManagerFileInfoImpl file) {
 		_file = file;
 	}
-	public int getOffset() {
+	public long getOffset() {
 		return _offset;
 	}
 	public int getLength() {
